pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[project]
name = "cuslines"
dynamic = ["version"]
description="GPU-accelerated tractography package"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "numpy",
    "nibabel",
    "tqdm",
    "dipy",
    "trx-python"
]

[project.optional-dependencies]
cu13 = [
    "nvidia-cuda-runtime",
    "nvidia-curand-cu12",
    "cuda-python<14",
    "cuda-core[cu13]",
    "cuda-cccl[cu13]"
]

cu12 = [
    "nvidia-cuda-runtime-cu12",
    "nvidia-curand-cu12",
    "cuda-python<13",
    "cuda-core[cu12]",
    "cuda-cccl[cu12]"
]

metal = [
    "pyobjc-framework-Metal",
    "pyobjc-framework-MetalPerformanceShaders",
]

webgpu = [
    "wgpu>=0.18",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cuslines*"]
